extension/WebInspectorKit/LayoutTests/inspector/debugger/debugger-script-preprocessor-expected.txt

CONSOLE MESSAGE: line 14: Uncaught Error: internal preprocessor error
CONSOLE MESSAGE: line 1: preprocessor threw exception
CONSOLE MESSAGE: line 1: Uncaught Error: failingPreprocessor throws
CONSOLE MESSAGE: line 1: Uncaught SyntaxError: Unexpected identifier
Tests script preprocessor (ability to preprocess all scripts upon reload).

Debugger was enabled.
Page reloaded.
sourceURLs[0]=inspector-test.js.js
sourceURLs[1]=debugger-test.js.js
sourceURLs[2]=debugger-script-preprocessor.html
sourceURLs[3]=debugger-script-preprocessor.html
sourceURLs[4]=dynamicScript1
sourceURLs[5]=dynamicScript2
sourceURLs[6]=
sourceURLs[7]=Bug226476.js
Page reloaded.
Page reloaded.
Page reloaded.
Debugger was disabled.
